refactor and separate create app form into components folder
